feat: migrate to nextjs
