+++ src/components/standings/data-table.tsx
@@ -1,0 +1,94 @@
+import React, { useState } from 'react';
+
+import {
+  flexRender,
+  getCoreRowModel,
+  useReactTable,
+  SortingState,
+  getSortedRowModel,
+} from '@tanstack/react-table';
+
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table';
+import { CustomColumnDef } from './columns';
+
+interface DataTableProps<TData, TValue> {
+  columns: CustomColumnDef<TData, TValue>[];
+  data: TData[];
+}
+
+export const DataTable = <TData, TValue>({ columns, data }: DataTableProps<TData, TValue>) => {
+  const [sorting, setSorting] = useState<SortingState>([]);
+
+  const table = useReactTable({
+    data,
+    columns,
+    getCoreRowModel: getCoreRowModel(),
+    onSortingChange: setSorting,
+    getSortedRowModel: getSortedRowModel(),
+    enableSortingRemoval: true,
+    state: {
+      sorting,
+    },
+  });
+
+  return (
+    <div className="rounded-md ">
+      <Table>
+        <TableHeader>
+          {table.getHeaderGroups().map((headerGroup) => (
+            <TableRow key={headerGroup.id}>
+              {headerGroup.headers.map((header) => {
+                const customHeaderClassName =
+                  (header.column.columnDef as CustomColumnDef<TData, TValue>)
+                    .customHeaderClassName || '';
+
+                return (
+                  <TableHead
+                    key={header.id}
+                    className={`px-0 text-redbull-grey ${customHeaderClassName}`}
+                  >
+                    {header.isPlaceholder
+                      ? null
+                      : flexRender(header.column.columnDef.header, header.getContext())}
+                  </TableHead>
+                );
+              })}
+            </TableRow>
+          ))}
+        </TableHeader>
+        <TableBody>
+          {table.getRowModel().rows?.length ? (
+            table.getRowModel().rows.map((row) => (
+              <TableRow key={row.id} data-state={row.getIsSelected() && 'selected'}>
+                {row.getVisibleCells().map((cell) => {
+                  const customCellClassName =
+                    (cell.column.columnDef as CustomColumnDef<TData, TValue>).customCellClassName ||
+                    '';
+
+                  return (
+                    <TableCell key={cell.id} className={`h-16 text-base ${customCellClassName}`}>
+                      {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                    </TableCell>
+                  );
+                })}
+              </TableRow>
+            ))
+          ) : (
+            <TableRow>
+              <TableCell colSpan={columns.length} className="h-24 text-center">
+                No results.
+              </TableCell>
+            </TableRow>
+          )}
+        </TableBody>
+      </Table>
+    </div>
+  );
+};
