+++ src/pages/index.tsx
@@ -1,0 +1,7 @@
+
+
+export default function Home() {
+  return (
+    <div> oi</div>
+  );
+}
