+++ src/pages/standings.tsx
@@ -1,0 +1,10 @@
+
+import StandingsTable from "@/components/standings";
+
+
+export default function Home() {
+  return (
+          <StandingsTable season={2023} />
+
+  );
+}
