--- test-01-driver-standings/src/App.css
@@ -1,13 +1,0 @@
-.App {
-  text-align: left;
-
-  @media screen and (min-width: 768px) {
-    padding: 4rem;
-    
-  }
-}
-
-
-
-
-
